Reorganizing the class to introduce git earlier
We may be able to get Patrick Masson here for the licensing lecture, so we will make that work to his schedule if possible. Either way, introducing the intricacies of git and distributed development earlier will give the class a better insight into how to develop in their project teams.