Rewrote the lab to use couchdb
I used a lot of material from the couchdb tutorial, but I also left the
mongodb stuff in (albeit commented out, I hope) for two reasons. The
first is in case we need to go back to Mongodb, we know it is there. The
second is that I would actually like to do this exercise using couchdb,
but I am wondering about how long this initial part will take and how
much I might have to cut out.